requirements.txt
pywin32>=305
python-pptx>=0.6.21
Pillow>=9.0.0
